listo

+++ Proyecto2Arbol/Proyecto2Arbol/Tools.cpp
@@ -13,7 +13,7 @@ int validaOpcion(std::string opcion){
 
 std::string direccion(std::string& valor) {
 	std::string direccion = "";
-	short tam = valor.size();
+	size_t tam = valor.size();
 	for (int i = 0; i < tam; i++) {
 		if (valor[i] == 'S' || valor[i] == 'N') {
 			direccion += valor[i];
